split cloudflare package
