+++ cloudflare/r2bucket.go
@@ -1,4 +1,4 @@
-package workers
+package cloudflare
 
 import (
 	"fmt"
