+++ cloudflare/r2objects.go
@@ -1,4 +1,4 @@
-package workers
+package cloudflare
 
 import (
 	"fmt"
